tests/wpt/web-platform-tests/trusted-types/createContextualFragment.tentative.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/helper.sub.js"></script>
<body>
<script>
  test(t => {
    var html = TrustedHTML.escape(STRINGS.unescapedHTML);

    var range = document.createRange();
    range.selectNodeContents(document.documentElement);
    var result = range.createContextualFragment(html);
    assert_equals(result.textContent, STRINGS.unescapedHTML);
  }, "DocumentFragment.textContent = TrustedHTML.escape().");

  test(t => {
    var html = TrustedHTML.unsafelyCreate(STRINGS.unescapedHTML);

    var range = document.createRange();
    range.selectNodeContents(document.documentElement);
    var result = range.createContextualFragment(html);
    assert_equals(result.textContent, STRINGS.unescapedText);
  }, "DocumentFragment.textContent = TrustedHTML.unsafelyCreate().");

  test(t => {
    var range = document.createRange();
    range.selectNodeContents(document.documentElement);
    var result = range.createContextualFragment(null);
    assert_equals(result.textContent, "null");
  }, "DocumentFragment.textContent = null.");
</script>
